requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.6.0
